% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_docker_resources_interactive.R
\name{clean_docker_resources_interactive}
\alias{clean_docker_resources_interactive}
\title{Clean Docker Resources Interactively}
\usage{
clean_docker_resources_interactive(cli = "docker")
}
\arguments{
\item{cli}{The CLI to use ("docker" or "podman"). Default is "docker".}
}
\description{
This function lists all Docker containers and images interactively, allowing the user to select which to remove.
}
