% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondorBox.R
\name{CondorBox}
\alias{CondorBox}
\title{Run Job via HTCondor and Docker with Make Options}
\usage{
CondorBox(
  remote_user,
  remote_host,
  remote_dir,
  github_pat,
  github_username,
  github_org,
  github_repo,
  stream_error = "FALSE",
  branch = "main",
  docker_image,
  target_folder = NULL,
  condor_cpus = NULL,
  condor_memory = NULL,
  condor_disk = NULL,
  make_options = "all",
  rmclone_script = "yes",
  ghcr_login = FALSE,
  remote_os = "linux",
  condor_environment = NULL,
  custom_batch_name = NULL
)
}
\arguments{
\item{remote_user}{Character. Remote server username.}

\item{remote_host}{Character. Remote server address.}

\item{remote_dir}{Character. Remote working directory.}

\item{github_pat}{Character. GitHub Personal Access Token.}

\item{github_username}{Character. GitHub username.}

\item{github_org}{Character. GitHub organisation name.}

\item{github_repo}{Character. GitHub repository name.}

\item{docker_image}{Character. Docker image to use.}

\item{target_folder}{Character, optional. Specific folder within the repository to clone (via sparse checkout)
and where \code{make} is executed. If not provided, the entire repository is used.}

\item{condor_cpus}{Numeric, optional. The number of CPUs to request in the HTCondor job.}

\item{condor_memory}{Character, optional. The amount of memory to request (e.g., "4GB") in the HTCondor job.}

\item{make_options}{Character, optional. Options to pass to the \code{make} command. Defaults to \code{"all"}.}
}
\value{
No return value; side effects include the creation and transfer of script files to the remote server
  and submission of an HTCondor job.
}
\description{
This function creates two Bash script files. The first script, \code{clone_job.sh},
handles cloning a GitHub repository (or a specific folder via sparse checkout) using the provided
GitHub Personal Access Token (PAT). The second script, \code{run_job.sh}, sources the clone script
to perform the clone, stores the working directory in a variable, schedules the deletion of the clone
script (to remove sensitive PAT information from disk), and then executes the remaining commands
(running \code{make} with user-defined options and archiving the appropriate folder). An HTCondor submit
file is also generated and transferred to a remote server for job submission.
}
