% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rserver_template.R
\name{generate_rserver_template}
\alias{generate_rserver_template}
\title{Generate an RStudio Server Docker Template}
\usage{
generate_rserver_template(output_file = "launch_rserver.R")
}
\arguments{
\item{output_file}{The name of the output file for the template script. Default is "launch_rserver.R".}
}
\value{
None. The function writes the template script to the specified file.
}
\description{
This function creates a script template for launching the RStudio server in a Docker container.
The generated script includes placeholders for the Docker image, container name, and other settings.
}
\examples{
\dontrun{
generate_rserver_template()
}
}
