% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_rstudio_server.R
\name{manage_rstudio_server}
\alias{manage_rstudio_server}
\title{Manage RStudio Server Containers}
\usage{
manage_rstudio_server(
  action = c("start", "stop"),
  image = "rocker/rstudio",
  container_name = "rstudio",
  host_port = 9999,
  container_port = 8787,
  password = "yourpassword"
)
}
\arguments{
\item{action}{A string indicating the action to perform: "start" or "stop".}

\item{image}{The Docker image to use. Default is "rocker/rstudio".}

\item{container_name}{The name of the container. Default is "rstudio".}

\item{host_port}{The port on the host machine. Default is 8787.}

\item{container_port}{The port inside the container. Default is 8787.}
}
\description{
This function starts or stops an RStudio Server container using Docker or Podman.
It supports creating new containers or restarting existing ones.
}
