% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_github_ssh.R
\name{setup_github_ssh}
\alias{setup_github_ssh}
\title{Setup GitHub SSH}
\usage{
setup_github_ssh(email, github_username, github_token = NULL)
}
\arguments{
\item{email}{Your GitHub email address.}

\item{github_username}{Your GitHub username.}

\item{github_token}{A GitHub personal access token with `write:public_key` scope (optional).}
}
\description{
This function sets up an SSH key for GitHub on RStudio Server, including generating a key,
adding it to the SSH agent, and optionally uploading it to GitHub via API.
}
