% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ssh_key.R
\name{setup_ssh_key}
\alias{setup_ssh_key}
\title{Setup SSH Key for Remote Server}
\usage{
setup_ssh_key(remote_user, remote_host)
}
\arguments{
\item{remote_user}{The username for the remote server.}

\item{remote_host}{The hostname or IP address of the remote server.}
}
\value{
None. Prints messages about the status of the setup.
}
\description{
This function generates an SSH key if it does not exist and copies the public key
to the specified remote server.
}
\examples{
\dontrun{
setup_ssh_key(remote_user = "your_username", remote_host = "your_remote_host")
}
}
